requirements.txt
pillow
accelerate
tqdm
transformers
datasets
